chore: seperate store and context
